requirements.txt
cryptography==46.0.1
pandas==2.3.2
python-dotenv==1.1.1
scikit_learn==1.7.2
snowflake_connector_python==3.15.0
